feat(mcp): extract registry tool metadata into registry-tools-lib
Move MCP public policy, read-only tool lists, and tool definition
metadata into a focused lib module with a thin registry re-export.
